Refactor test patching and do it on load
