+++ src/services/import_parsing.ts
@@ -1,7 +1,4 @@
-import { Base64 } from "js-base64"
-
 import { FileEntry } from "../components/ProgrammingExerciseLoader"
-import { patchTmcUtilsPy } from "./test_patching"
 
 type PythonImportAll = {
   pkg: string
@@ -26,74 +23,6 @@ const resolveImports = (start: FileEntry, files: FileEntry[]): string => {
   return wrap(start.content, [start.shortName], files)
 }
 
-const resolveTestImports = (
-  code: string,
-  test: FileEntry,
-  tmcFiles: FileEntry[],
-): string => {
-  const files = tmcFiles.map((x) => {
-    let content = removeRelativeTmcImports(x.content)
-    switch (x.shortName) {
-      case "utils.py":
-        return { ...x, content: patchTmcUtilsPy(content) }
-      default:
-        return { ...x, content }
-    }
-  })
-
-  const tempHelpTmcEncode = (fileName: string) => {
-    const file = files.find((x) => x.shortName === fileName) as FileEntry
-    const wrapped = wrap(file.content, [file.shortName], tmcFiles)
-    console.log(fileName, wrapped)
-    return Base64.encode(wrapped)
-  }
-
-  const template = `
-import sys, base64, importlib
-from types import ModuleType
-
-_stdout_pointer = 0
-
-def __decode(code_b64):
-    return base64.b64decode(code_b64).decode("utf-8")
-
-def __wrap_import(module_name, code):
-    mod = ModuleType(module_name)
-    sys.modules[module_name] = mod
-    exec(code, mod.__dict__)
-
-__wrap_import("webeditor", """
-import base64
-code = base64.b64decode("${Base64.encode(code)}").decode("utf-8")
-""")
-
-__wrap_import("tmc_points", __decode("${tempHelpTmcEncode("points.py")}"))
-__wrap_import("tmc_result", __decode("${tempHelpTmcEncode("result.py")}"))
-__wrap_import("tmc_runner", __decode("${tempHelpTmcEncode("runner.py")}"))
-__wrap_import("tmc_utils", __decode("${tempHelpTmcEncode("utils.py")}"))
-
-${test.content.replaceAll(
-  /\w+Test\(unittest.TestCase\)/g,
-  "PythonEditorTest(unittest.TestCase)",
-)}
-
-testOutput = ""
-from tmc.runner import TMCTestRunner
-from tmc.result import results
-
-import io, json, contextlib
-from unittest import TextTestRunner
-test_suite = unittest.TestLoader().loadTestsFromTestCase(PythonEditorTest)
-with io.StringIO() as buf:
-    with contextlib.redirect_stdout(buf):
-        TMCTestRunner(stream=buf).run(test_suite)
-    testOutput = json.dumps(results, ensure_ascii=False)
-`
-  console.log(template)
-  const wrapped = wrap(template, [test.shortName], files)
-  return wrapped
-}
-
 const wrap = (
   source: string,
   presentlyImported: Array<string>,
@@ -101,7 +30,6 @@ const wrap = (
 ) => {
   const importAllPattern = /^import \./
   const importSomePattern = /^from \.\w+ import/
-  const importTmcPattern = /^from tmc\.?[a-zA-Z0-9]* import/
   const sourceLines = source.split("\n")
   const lines = sourceLines.map((line, num) => {
     if (line.match(importAllPattern)) {
@@ -115,31 +43,9 @@ const wrap = (
         files,
       )
     }
-    if (line.match(importTmcPattern)) {
-      return replaceImportTmc(line)
-    }
     return line
   })
   return lines.join("\n")
-}
-
-const removeRelativeTmcImports = (source: string): string => {
-  const lines = source.split("\n")
-  return lines
-    .map((line) => {
-      let match = line.match(/^import \.(\w+)/)
-      if (match) {
-        return `import tmc_${match[1]}`
-      }
-
-      match = line.match(/^from \.(\w+) import ([\w,\s]+)/)
-      if (match) {
-        return `from tmc_${match[1]} import ${match[2]}`
-      }
-
-      return line
-    })
-    .join("\n")
 }
 
 const replaceImportAll = (
@@ -224,21 +130,4 @@ const getContentByShortName = (name: string, fileSet: Array<FileEntry>) => {
   return fileSet.filter(({ shortName }) => shortName === name)[0].content
 }
 
-const replaceImportTmc = (line: string): string => {
-  const importMatches = line.match(/^from tmc import ([\w,\s]+)/)
-  if (importMatches) {
-    return importMatches[1]
-      .split(",")
-      .map((pkg) => `from tmc_${pkg.trim()} import ${pkg.trim()}`)
-      .join("\n")
-  }
-
-  const importMatches2 = line.match(/from tmc\.(\w+) import ([\w,\s]+)/)
-  if (importMatches2) {
-    return `from tmc_${importMatches2[1]} import ${importMatches2[2]}`
-  }
-
-  throw `Malformed import statement: ${line}`
-}
-
-export { resolveImports, resolveTestImports }
+export { resolveImports }
